+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 ##1.8.33**
+* Add helper endpoint in dev mode to get Java System Properties #1320
 * FileBinder - When running command from CLI over HTTP, bind to new File #1319
 * Allow app to fetch configuration from a configuration server #1318
 * skip loading "META-INF/versions/9/module-info" class #1288